+++ benchmarks/transport/CMakeLists.txt
@@ -1,8 +1,10 @@
 # Variable used for benchmarks that DO NOT require multithreading support
-set(_benchmarks ghex_p2p_bi_cb_avail ghex_p2p_bi_cb_wait ghex_p2p_bi_ft_avail ghex_p2p_bi_ft_wait mpi_p2p_bi_avail mpi_p2p_bi_wait)
+#set(_benchmarks ghex_p2p_bi_cb_avail ghex_p2p_bi_cb_wait ghex_p2p_bi_ft_avail ghex_p2p_bi_ft_wait mpi_p2p_bi_avail mpi_p2p_bi_wait)
+set(_benchmarks ghex_p2p_bi_cb_avail)
 
 # Variable used for benchmarks that require multithreading support
-set(_benchmarks_mt ghex_p2p_bi_cb_avail ghex_p2p_bi_cb_wait ghex_p2p_bi_ft_avail ghex_p2p_bi_ft_wait mpi_p2p_bi_avail mpi_p2p_bi_wait)
+#set(_benchmarks_mt ghex_p2p_bi_cb_avail ghex_p2p_bi_cb_wait ghex_p2p_bi_ft_avail ghex_p2p_bi_ft_wait mpi_p2p_bi_avail mpi_p2p_bi_wait)
+set(_benchmarks_mt ghex_p2p_bi_cb_avail)
 
 foreach (_t ${_benchmarks})
     add_executable(${_t} ${_t}_mt.cpp )
