+++ benchmarks/transport/ghex_p2p_bi_cb_avail_mt_new.cpp
@@ -1,8 +1,6 @@
 #include <iostream>
 #include <vector>
 #include <atomic>
-
-#include <ghex/transport_layer/ucx/threads.hpp>
 #include <ghex/common/timer.hpp>
 #include "utils.hpp"
 
@@ -30,7 +28,7 @@ namespace ghex = gridtools::ghex;
         #include <ghex/threads/none/primitives.hpp>
         using threading    = ghex::threads::none::primitives;
     #endif
-    #include <ghex/transport_layer/ucx3/address_db_mpi.hpp>
+    #include <ghex/transport_layer/ucx/address_db_mpi.hpp>
     #include <ghex/transport_layer/ucx/context.hpp>
     using db_type      = ghex::tl::ucx::address_db_mpi;
     using transport    = ghex::tl::ucx_tag;
@@ -46,7 +44,6 @@ using MsgType = gridtools::ghex::tl::message_buffer<>;
 //using MsgType = gridtools::ghex::tl::shared_message_buffer<>;
 
 
-/* comm requests currently in-flight */
 std::atomic<int> sent(0);
 std::atomic<int> received(0);
 std::atomic<int> tail_send(0);
@@ -54,40 +51,6 @@ std::atomic<int> tail_recv(0);
 int last_received = 0;
 int last_sent = 0;
 int inflight;
-
-/* Track finished comm requests. 
-   This is shared between threads, because in the shared-worker case
-   there is no way of knowing which thread will service which requests,
-   and how many.
-*/
-int comm_cnt = 0, nlsend_cnt = 0, nlrecv_cnt = 0, submit_cnt = 0, submit_recv_cnt = 0;
-DECLARE_THREAD_PRIVATE(comm_cnt)
-DECLARE_THREAD_PRIVATE(nlsend_cnt)
-DECLARE_THREAD_PRIVATE(nlrecv_cnt)
-DECLARE_THREAD_PRIVATE(submit_cnt)
-DECLARE_THREAD_PRIVATE(submit_recv_cnt)
-
-int thrid, nthr;
-DECLARE_THREAD_PRIVATE(thrid)
-DECLARE_THREAD_PRIVATE(nthr)
-
-void send_callback(communicator_type::message_type, int, int tag)
-{
-    // std::cout << "send callback called " << /*rank <<*/ /*" thread " << omp_get_thread_num() <<*/ " tag " << tag << "\n";
-    //int pthr = tag/inflight;
-    //if(pthr != thrid) nlsend_cnt++;
-    //comm_cnt++;
-    sent++;
-}
-
-void recv_callback(communicator_type::message_type, int, int tag)
-{
-     //std::cout << "recv callback called " << /*rank <<*/ /*" thread " << omp_get_thread_num() <<*/ " tag " << tag /*<< " ongoing " << ongoing_comm*/ << "\n";
-    //int pthr = tag/inflight;
-    //if(pthr != thrid) nlrecv_cnt++;
-    //comm_cnt++;
-    received++;
-}
 
 int main(int argc, char *argv[])
 {
@@ -127,38 +90,41 @@ int main(int argc, char *argv[])
         context_type context{num_threads, MPI_COMM_WORLD, db_type{MPI_COMM_WORLD} };
         #endif
 
-
-      THREAD_PARALLEL_BEG() {
-
-            gridtools::ghex::timer timer_test;
-            gridtools::ghex::timer timer_send;
-            gridtools::ghex::timer timer_recv;
-            gridtools::ghex::timer timer_prog;
-            gridtools::ghex::timer timer_while;
-            gridtools::ghex::timer timer_while2;
-            gridtools::ghex::timer timer_body;
-            gridtools::ghex::timer timer_cond;
-
-            auto token = context.get_token();
-
-            auto comm = context.get_communicator(token);
-
-            const auto rank = comm.rank();
-            const auto size = comm.size();
-            const auto thread_id = token.id();
+#ifdef USE_OPENMP
+        #pragma omp parallel
+#endif
+        {
+            auto token             = context.get_token();
+            auto comm              = context.get_communicator(token);
+            const auto rank        = comm.rank();
+            const auto size        = comm.size();
+            const auto thread_id   = token.id();
             const auto num_threads = context.thread_primitives().size();
-            const auto peer_rank = (rank+1)%2;
-            thrid = token.id();
-            nthr = num_threads;
+            const auto peer_rank   = (rank+1)%2;
             
-            //context.thread_primitives().master(token, 
-            //   [rank,&comm]()
-                if (token.id()==0)
-                {
-//	THREAD_MASTER() {
-                    if(rank==0)
-                    std::cout << "\n\nrunning test " << __FILE__ << " with communicator " << typeid(comm).name() << "\n\n";
-                };
+            int comm_cnt = 0, nlsend_cnt = 0, nlrecv_cnt = 0, submit_cnt = 0, submit_recv_cnt = 0;
+            int dbg = 0, sdbg = 0, rdbg = 0;
+
+            auto send_callback = [&](communicator_type::message_type, int, int tag)
+            {
+                int pthr = tag/inflight;
+                if(pthr != thread_id) nlsend_cnt++;
+                comm_cnt++;
+                sent++;
+            };
+
+            auto recv_callback = [&](communicator_type::message_type, int, int tag)
+            {
+                int pthr = tag/inflight;
+                if(pthr != thread_id) nlrecv_cnt++;
+                comm_cnt++;
+                received++;
+            };
+            
+            if (thread_id==0 && rank==0)
+            {
+                std::cout << "\n\nrunning test " << __FILE__ << " with communicator " << typeid(comm).name() << "\n\n";
+            };
             
             std::vector<MsgType> smsgs(inflight);
             std::vector<MsgType> rmsgs(inflight);
@@ -173,35 +139,20 @@ int main(int argc, char *argv[])
             }
             
             context.barrier(token);
-  //          context.world().barrier();
             
-            //context.thread_primitives().master(token, 
-            //    [&timer,&ttimer,rank,num_threads]() 
-//                #pragma omp master
-                if (token.id()==0)
-                { 
-                    timer.tic();
-                    ttimer.tic();
-                    if(rank == 1) 
-                        std::cout << "number of threads: " << num_threads << ", multi-threaded: " << THREAD_IS_MT << "\n";
-                };
-//	THREAD_BARRIER();
-
-            //context.thread_primitives().barrier(token);
+            if (thread_id == 0)
+            { 
+                timer.tic();
+                ttimer.tic();
+                if(rank == 1) 
+                    std::cout << "number of threads: " << num_threads << ", multi-threaded: true\n";
+            };
             
             // send/recv niter messages - as soon as a slot becomes free
-            int dbg = 0, sdbg = 0, rdbg = 0;
-            bool loop_condition = sent < niter || received < niter;
+            while(sent < niter || received < niter)
             {
-            timer_while.tic();
-            std::cout << "start\n";
-            timer_while.toc();
-            timer_while.tic();
-            timer_while2.tic();
-            while(true)
-            {
-                //timer_body.tic();
-                if(thrid == 0 && dbg >= (niter/10)) {
+                if(thread_id == 0 && dbg >= (niter/10))
+                {
                     dbg = 0;
                     std::cout << rank << " total bwdt MB/s:      " 
                               << ((double)(received-last_received + sent-last_sent)*size*buff_size/2)/timer.stoc()
@@ -211,111 +162,60 @@ int main(int argc, char *argv[])
                     last_sent = sent;
                 }
 
-                if(rank==0 && thrid==0 && rdbg >= (niter/10)) {
+                if(rank==0 && thread_id==0 && rdbg >= (niter/10))
+                {
                     std::cout << received << " received\n";
                     rdbg = 0;
                 }
 
-                if(rank==0 && thrid==0 && sdbg >= (niter/10)) {
+                if(rank==0 && thread_id==0 && sdbg >= (niter/10))
+                {
                     std::cout << sent << " sent\n";
                     sdbg = 0;
                 }
-                for(int j=0; j<inflight; j++){
-                    //if(rmsgs[j].use_count() == 1){
-                    bool r_test;
-                 //   timer_test.tic();
-                    //r_test = rmsgs[j].use_count()==1;
-                    r_test = rreqs[j].test();
-               //     timer_test.toc();
-                    //if (rreqs[j].test()) {
-                    if (r_test) {
-                        submit_recv_cnt += nthr;
-                        rdbg += nthr;
-                        dbg += nthr;
-                       // timer_recv.tic();
-                        //auto msg = rmsgs[j];
-                        //rreqs[j] = comm.recv(std::move(msg), peer_rank, thrid*inflight+j, recv_callback);
-                        rreqs[j] = comm.recv(rmsgs[j], peer_rank, thrid*inflight+j, recv_callback);
-                     //   timer_recv.toc();
-                    } else 
-                    {
-                   //     timer_prog.tic();
-                        comm.progress();
-                   //     timer_prog.toc();
-                    }
 
-                    //if(sent < niter && smsgs[j].use_count() == 1){
-                    bool s_test;
-                 //   timer_test.tic();
-                    //s_test = smsgs[j].use_count()==1;
-                    s_test = sreqs[j].test();
-                 //   timer_test.toc();
-                    //if(sent < niter && sreqs[j].test()){
-                    if(sent < niter && s_test){
-                        submit_cnt += nthr;
-                        sdbg += nthr;
-                        dbg += nthr;
-                      //  timer_send.tic();
-                        //auto msg = smsgs[j];
-                        //sreqs[j] = comm.send(std::move(msg), peer_rank, thrid*inflight+j, send_callback);
-                        sreqs[j] = comm.send(smsgs[j], peer_rank, thrid*inflight+j, send_callback);
-                     //   timer_send.toc();
-                    } else
-                    {
-                   //     timer_prog.tic();
-                        comm.progress();
-                  //      timer_prog.toc();
-                    }
-                }
-                //timer_body.toc();
-                //timer_cond.tic();
-                loop_condition = sent < niter || received < niter;
-                if (!loop_condition) break;
-                //timer_cond.toc();
-                //timer_while2.toc_tic();
-            }
-            //timer_while2.toc();
-            timer_while.toc();
-            }
-                /*std::cout << "test  " << timer_test.sum() << std::endl;
-                std::cout << "send  " << timer_send.sum() << std::endl;
-                std::cout << "recv  " << timer_recv.sum() << std::endl;
-                std::cout << "prog  " << timer_prog.sum() << std::endl;
-                std::cout << "loop  " << timer_while.sum() << std::endl;
-                std::cout << "loop2 " << timer_while2.sum() << std::endl;
-                std::cout << "body  " << timer_body.sum() << std::endl;
-                std::cout << "cond  " << timer_cond.sum() << std::endl;*/
-            
-//	THREAD_BARRIER()
-//	THREAD_MASTER() {
-	    MPI_Barrier(MPI_COMM_WORLD);
-	    if(rank == 1) {
-		ttimer.vtoc();
-		ttimer.vtoc("final ", (double)niter*size*buff_size);
-//	    }
-	}
-//	THREAD_BARRIER()
-    /*        context.barrier(token);
-            context.thread_primitives().master(token, 
-                [&niter,size,buff_size,&ttimer,rank,num_threads]() 
-                { 
-                    if (rank==1)
-                    {
-                        const auto t = ttimer.stoc();
-                        std::cout << "time:       " << t/1000000 << "s\n";
-                        std::cout << "final MB/s: " << ((double)niter*size*buff_size)/t << "\n";
-                    }       
-                });
-            
-            context.thread_primitives().barrier(token);*/
-
-            //context.thread_primitives().critical(
-             //   [&]()
+                for(int j=0; j<inflight; j++)
                 {
-                    std::cout << "rank " << rank << " thread " << thrid << " sends submitted " << submit_cnt/nthr
-                              << " serviced " << comm_cnt << ", non-local sends " << nlsend_cnt << " non-local recvs " << nlrecv_cnt << "\n";
+                    //if(rmsgs[j].use_count() == 1)
+                    if (rreqs[j].test())
+                    {
+                        submit_recv_cnt += num_threads;
+                        rdbg += num_threads;
+                        dbg += num_threads;
+                        rreqs[j] = comm.recv(rmsgs[j], peer_rank, thread_id*inflight+j, recv_callback);
+                    }
+                    else 
+                        comm.progress();
 
-                }//);
+                    //if(sent < niter && smsgs[j].use_count() == 1)
+                    if(sent < niter && sreqs[j].test())
+                    {
+                        submit_cnt += num_threads;
+                        sdbg += num_threads;
+                        dbg += num_threads;
+                        sreqs[j] = comm.send(smsgs[j], peer_rank, thread_id*inflight+j, send_callback);
+                    }
+                    else
+                        comm.progress();
+                }
+            }
+	        
+            if(thread_id==0 && rank == 0)
+            {
+                const auto t = ttimer.stoc();
+                std::cout << "time:       " << t/1000000 << "s\n";
+                std::cout << "final MB/s: " << ((double)niter*size*buff_size)/t << "\n";
+	        }
+
+            context.barrier(token);
+
+            context.thread_primitives().critical(
+            [&]()
+            {
+                std::cout 
+                << "rank " << rank << " thread " << thread_id << " sends submitted " << submit_cnt/num_threads
+                << " serviced " << comm_cnt << ", non-local sends " << nlsend_cnt << " non-local recvs " << nlrecv_cnt << "\n";
+            });
 
             // tail loops - submit RECV requests until
             // all SEND requests have been finalized.
@@ -344,7 +244,6 @@ int main(int argc, char *argv[])
                     // continue to re-schedule all recvs to allow the peer to complete
                     for(int j=0; j<inflight; j++){
                         if(rreqs[j].test()) {
-                            //auto msg = rmsgs[j];
                             rreqs[j] = comm.recv(rmsgs[j], peer_rank, thread_id*inflight + j, recv_callback);
                         }
                     }
@@ -354,12 +253,12 @@ int main(int argc, char *argv[])
                 // Notify the peer and keep submitting recvs until we get his notification.
                 future_type sf, rf;
                 MsgType smsg(1), rmsg(1);
-                //context.thread_primitives().master(token, 
-                //    [&comm,&sf,&rf,&smsg,&rmsg,peer_rank]() mutable
+                context.thread_primitives().master(token, 
+                    [&]() mutable
                     {
                         sf = comm.send(smsg, peer_rank, 0x800000, [](communicator_type::message_type, int, int){});
                         rf = comm.recv(rmsg, peer_rank, 0x800000, [](communicator_type::message_type, int, int){});
-                    }//);
+                    });
 
                 while(!tail_recv.load()){
                     comm.progress();
@@ -367,23 +266,21 @@ int main(int argc, char *argv[])
                     // schedule all recvs to allow the peer to complete
                     for(int j=0; j<inflight; j++){
                         if(rreqs[j].test()) {
-                            //auto msg = rmsgs[j];
                             rreqs[j] = comm.recv(rmsgs[j], peer_rank, thread_id*inflight + j, recv_callback);
                         }
                     }
-                   // context.thread_primitives().master(token, 
-                    //    [&rf,&tail_recv]()
+                    context.thread_primitives().master(token, 
+                        [&]()
                         {
                             if(rf.test()) tail_recv = 1;
-                        }//);
+                        });
                 }
             }
             // peer has sent everything, so we can cancel all posted recv requests
             for(int j=0; j<inflight; j++){
                 rreqs[j].cancel();
             }
-
-        } THREAD_PARALLEL_END();    
+        }
     }
 
     MPI_Barrier(MPI_COMM_WORLD);
