+++ include/ghex/transport_layer/ucx/future.hpp
@@ -19,23 +19,23 @@ namespace gridtools{
             namespace ucx {
 
                 /** @brief future template for non-blocking communication */
-                template<typename T>
-                struct future
+                template<typename T, typename ThreadPrimitives>
+                struct future_t
                 {
                     using value_type  = T;
-                    using handle_type = request;
+                    using handle_type = request_ft<ThreadPrimitives>;
 
                     value_type m_data;
                     handle_type m_handle;
 
-                    future(value_type&& data, handle_type&& h) 
+                    future_t(value_type&& data, handle_type&& h) 
                     :   m_data(std::move(data))
                     ,   m_handle(std::move(h))
                     {}
-                    future(const future&) = delete;
-                    future(future&&) = default;
-                    future& operator=(const future&) = delete;
-                    future& operator=(future&&) = default;
+                    future_t(const future_t&) = delete;
+                    future_t(future_t&&) = default;
+                    future_t& operator=(const future_t&) = delete;
+                    future_t& operator=(future_t&&) = default;
 
                     void wait() noexcept
                     {
@@ -62,25 +62,25 @@ namespace gridtools{
                       * @return True if the request was successfully canceled */
                     bool cancel()
                     {
-			return m_handle.cancel();
+                        return m_handle.cancel();
                     }
                 };
 
-                template<>
-                struct future<void>
+                template<typename ThreadPrimitives>
+                struct future_t<void, ThreadPrimitives>
                 {
-                    using handle_type = request;
+                    using handle_type = request_ft<ThreadPrimitives>;
 
                     handle_type m_handle;
 
-                    future() noexcept = default; 
-                    future(handle_type&& h) 
+                    future_t() noexcept = default; 
+                    future_t(handle_type&& h) 
                     :   m_handle(std::move(h))
                     {}
-                    future(const future&) = delete;
-                    future(future&&) = default;
-                    future& operator=(const future&) = delete;
-                    future& operator=(future&&) = default;
+                    future_t(const future_t&) = delete;
+                    future_t(future_t&&) = default;
+                    future_t& operator=(const future_t&) = delete;
+                    future_t& operator=(future_t&&) = default;
 
                     void wait() noexcept
                     {
@@ -104,57 +104,7 @@ namespace gridtools{
 
                     bool cancel()
                     {
-			return m_handle.cancel();
-                    }
-                };
-
-		template<typename Allocator>
-                struct future_cb
-                {
-                    using handle_type = request_cb<Allocator>;
-
-                    handle_type m_handle;
-
-                    future_cb() {}
-                    future_cb(handle_type&& h):
-			m_handle(std::move(h))
-                    {
-			fprintf(stderr, "%s %d m_req %p\n", __FUNCTION__, __LINE__, m_handle.m_req);
-		    }
-
-                    future_cb(const future_cb&) = delete;
-                    future_cb(future_cb&& other) = default;
-                    future_cb& operator=(const future_cb&) = delete;
-                    future_cb& operator=(future_cb&& other)
-		    {
-		    	fprintf(stderr, "%s %d m_req %p\n", __FUNCTION__, __LINE__, other.m_handle.m_req);
-		    	m_handle = std::move(other.m_handle);
-		    	fprintf(stderr, "%s %d m_req %p\n", __FUNCTION__, __LINE__, m_handle.m_req);
-		    }
-
-                    void wait() noexcept
-                    {
-                        m_handle.wait();
-                    }
-
-                    bool test() noexcept
-                    {
-                        return m_handle.test();
-                    }
-
-                    bool ready() noexcept
-                    {
-                        return m_handle.test();
-                    }
-
-                    void get()
-                    {
-                        wait(); 
-                    }
-
-                    bool cancel()
-                    {
-			return m_handle.cancel();
+                        return m_handle.cancel();
                     }
                 };
