+++ tests/transport/test_ucx_context.cpp
@@ -9,7 +9,7 @@
  * 
  */
 #include <iostream>
-#include <ghex/transport_layer/ucx3/address_db_mpi.hpp>
+#include <ghex/transport_layer/ucx/address_db_mpi.hpp>
 #include <ghex/transport_layer/ucx/context.hpp>
 #include <ghex/threads/atomic/primitives.hpp>
